requirements.txt
Flask==2.1.2
pywebview~=4.3.3
requests~=2.31.0